Move testing utilities to `src/utilities/testing`
Since we're combining all core + React testing utilities, let's
get them out of the `__tests__` dir. It seemed a bit strange to
create a separate `testing` bundle from a `__tests__` source
location.
