+++ src/__tests__/ApolloClient.ts
@@ -5,8 +5,8 @@ import { makeReference } from '../utilities/graphql/storeUtils';
 import { ApolloLink } from '../link/core/ApolloLink';
 import { HttpLink } from '../link/http/HttpLink';
 import { InMemoryCache } from '../cache/inmemory/inMemoryCache';
-import { stripSymbols } from './utils/stripSymbols';
-import { withWarning } from './utils/wrap';
+import { stripSymbols } from '../utilities/testing/stripSymbols';
+import { withWarning } from '../utilities/testing/wrap';
 import { ApolloClient } from '../';
 import { DefaultOptions } from '../ApolloClient';
 import { FetchPolicy, QueryOptions } from '../core/watchQueryOptions';
