+++ src/__tests__/fetchMore.ts
@@ -1,10 +1,10 @@
 import { assign, cloneDeep } from 'lodash';
 import gql from 'graphql-tag';
 
-import { mockSingleLink } from './utils/mocks/mockLink';
+import { mockSingleLink } from '../utilities/testing/mocks/mockLink';
 import { InMemoryCache } from '../cache/inmemory/inMemoryCache';
 import { ApolloClient, NetworkStatus, ObservableQuery } from '../';
-import { itAsync } from './utils/itAsync';
+import { itAsync } from '../utilities/testing/itAsync';
 
 describe('updateQuery on a simple query', () => {
   const query = gql`
