+++ src/__tests__/local-state/general.ts
@@ -9,7 +9,7 @@ import { ApolloClient } from '../..';
 import { ApolloCache } from '../../cache/core/cache';
 import { InMemoryCache } from '../../cache/inmemory/inMemoryCache';
 import { hasDirectives } from '../../utilities/graphql/directives';
-import { itAsync } from '../utils/itAsync';
+import { itAsync } from '../../utilities/testing/itAsync';
 
 describe('General functionality', () => {
   it('should not impact normal non-@client use', () => {
