+++ src/__tests__/local-state/resolvers.ts
@@ -5,9 +5,9 @@ import { assign } from 'lodash';
 import { Observable, Observer } from '../../utilities/observables/Observable';
 import { ApolloLink } from '../../link/core/ApolloLink';
 import { ApolloClient } from '../..';
-import mockQueryManager from '../utils/mocks/mockQueryManager';
-import wrap from '../../__tests__/utils/wrap';
-import { itAsync } from '../utils/itAsync';
+import mockQueryManager from '../../utilities/testing/mocks/mockQueryManager';
+import wrap from '../../utilities/testing/wrap';
+import { itAsync } from '../../utilities/testing/itAsync';
 import { ApolloQueryResult, Resolvers } from '../../core/types';
 import { WatchQueryOptions } from '../../core/watchQueryOptions';
 import { LocalState } from '../../core/LocalState';
