+++ src/cache/inmemory/__tests__/cache.ts
@@ -1,6 +1,6 @@
 import gql, { disableFragmentWarnings } from 'graphql-tag';
 
-import { stripSymbols } from '../../../__tests__/utils/stripSymbols';
+import { stripSymbols } from '../../../utilities/testing/stripSymbols';
 import { cloneDeep } from '../../../utilities/common/cloneDeep';
 import { makeReference } from '../../../utilities/graphql/storeUtils';
 import { InMemoryCache, InMemoryCacheConfig } from '../inMemoryCache';
