+++ src/core/__tests__/QueryManager/links.ts
@@ -4,10 +4,10 @@ import gql from 'graphql-tag';
 import { Observable } from '../../../utilities/observables/Observable';
 import { ApolloLink } from '../../../link/core/ApolloLink';
 import { InMemoryCache } from '../../../cache/inmemory/inMemoryCache';
-import { stripSymbols } from '../../../__tests__/utils/stripSymbols';
+import { stripSymbols } from '../../../utilities/testing/stripSymbols';
 
 // mocks
-import { MockSubscriptionLink } from '../../../__tests__/utils/mocks/mockSubscriptionLink';
+import { MockSubscriptionLink } from '../../../utilities/testing/mocks/mockSubscriptionLink';
 
 // core
 import { QueryManager } from '../../QueryManager';
