+++ src/core/__tests__/QueryManager/multiple-results.ts
@@ -1,12 +1,12 @@
 // externals
 import gql from 'graphql-tag';
 import { InMemoryCache } from '../../../cache/inmemory/inMemoryCache';
-import { stripSymbols } from '../../../__tests__/utils/stripSymbols';
+import { stripSymbols } from '../../../utilities/testing/stripSymbols';
 
 // mocks
 import {
   MockSubscriptionLink
-} from '../../../__tests__/utils/mocks/mockSubscriptionLink';
+} from '../../../utilities/testing/mocks/mockSubscriptionLink';
 
 // core
 import { QueryManager } from '../../QueryManager';
