+++ src/core/__tests__/QueryManager/recycler.ts
@@ -7,10 +7,10 @@
 // externals
 import gql from 'graphql-tag';
 import { InMemoryCache } from '../../../cache/inmemory/inMemoryCache';
-import { stripSymbols } from '../../../__tests__/utils/stripSymbols';
+import { stripSymbols } from '../../../utilities/testing/stripSymbols';
 import {
   MockSubscriptionLink
-} from '../../../__tests__/utils/mocks/mockSubscriptionLink';
+} from '../../../utilities/testing/mocks/mockSubscriptionLink';
 
 // core
 import { QueryManager } from '../../QueryManager';
