+++ src/react/hooks/__tests__/useMutation.test.tsx
@@ -4,7 +4,7 @@ import gql from 'graphql-tag';
 import { render, cleanup, wait } from '@testing-library/react';
 
 import { MockedProvider, mockSingleLink } from '../../testing';
-import { itAsync } from '../../../__tests__/utils/itAsync';
+import { itAsync } from '../../../utilities/testing/itAsync';
 import { ApolloClient } from '../../../ApolloClient';
 import { InMemoryCache } from '../../../cache/inmemory/inMemoryCache';
 import { ApolloProvider } from '../../context/ApolloProvider';
