+++ src/react/hooks/__tests__/useQuery.test.tsx
@@ -6,8 +6,8 @@ import { render, cleanup, wait } from '@testing-library/react';
 import { Observable } from '../../../utilities/observables/Observable';
 import { ApolloLink } from '../../../link/core/ApolloLink';
 import { MockedProvider } from '../../testing';
-import { MockLink } from '../../../__tests__/utils/mocks/mockLink';
-import { itAsync } from '../../../__tests__/utils/itAsync';
+import { MockLink } from '../../../utilities/testing/mocks/mockLink';
+import { itAsync } from '../../../utilities/testing/itAsync';
 import { ApolloClient } from '../../../ApolloClient';
 import { InMemoryCache } from '../../../cache/inmemory/inMemoryCache';
 import { ApolloProvider } from '../../context/ApolloProvider';
