+++ src/utilities/testing/stripSymbols.ts
