+++ tsconfig.json
@@ -19,6 +19,7 @@
   },
   "include": ["src/**/*.ts"],
   "exclude": [
-    "src/**/__tests__/**/*"
+    "src/**/__tests__/**/*",
+    "src/utilities/testing/**/*"
   ]
 }
